Removes duplicate ImageDecryptor, IImageDecryptor
Removes duplicate ImageDecryptor and IImageDecryptor and moves
implementation to where it is expected to be by controllers -
Service.Decryptor. ImageDecryptorTest restored to initial version 5edb5eab
