+++ Comptech.Backend/src/Comptech.Backend.Service/Decryptor/IImageDecryptor.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
+﻿
 namespace Comptech.Backend.Service.Decryptor
 {
     public interface IImageDecryptor
