+++ Comptech.Backend/src/Comptech.Backend.Service/Decryptor/libEncrypt.so
