--- Comptech.Backend/src/Comptech.Backend.Service/IImageDecryptor.cs
@@ -1,7 +1,0 @@
-﻿namespace Comptech.Backend.Service
-{
-    public interface IImageDecryptor
-    {
-        byte[] Decrypt(byte[] encryptedPhoto);
-    }
-}
